[relayer]: extract withdrawal module from pallet lib.rs
Moves `Pallet::withdraw` and the free `message` helper out of lib.rs into
withdrawal.rs alongside the types they already share. The pallet macro
shell stays in lib.rs; the withdrawal flow now lives in one file.

Drops the dead `FeeMetadata` RLP struct from withdrawal.rs in passing
(zero callers across modules/ tesseract/ sdk/).

Re-exports `pub use withdrawal::message;` at the crate root so the
existing `pallet_ismp_relayer::message` import in the testsuite keeps
working.
