+++ modules/pallets/relayer/src/withdrawal.rs
@@ -13,14 +13,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//! Relayer fee withdrawal.
+//!
+//! Once fees have been accumulated into [`crate::pallet::Fees`] by
+//! [`crate::accumulate`], relayers withdraw them via [`Pallet::withdraw`].
+//! The flow:
+//!
+//! 1. The relayer signs a `(nonce, dest_chain, beneficiary?)` payload with their per-chain
+//!    key (EVM secp256k1 / sr25519 / ed25519).
+//! 2. The pallet verifies the signature, increments the per-relayer nonce, and dispatches an
+//!    ISMP POST request to the destination's host manager (EVM) or `HYPERBRIDGE_MODULE_ID`
+//!    (substrate) instructing it to disburse `available_amount` of the fee token to the
+//!    beneficiary.
+//! 3. The `Fees` entry is zeroed so the same balance cannot be withdrawn twice.
+//!
+//! The on-chain effect is just dispatching the message; the destination chain settles the
+//! payout when the ISMP request is delivered there.
+
+use crate::{Config, Error, Event, Fees, MinWithdrawal, Nonce, Pallet, MODULE_ID};
 use alloc::vec::Vec;
-use alloy_primitives::Address;
-use alloy_rlp_derive::{RlpDecodable, RlpEncodable};
 use codec::{Decode, DecodeWithMemTracking, Encode};
 pub use crypto_utils::verification::Signature;
-use ismp::{host::StateMachine, messaging::Proof};
+use frame_support::{dispatch::DispatchResult, traits::Get};
+use ismp::{
+	dispatcher::{DispatchPost, DispatchRequest, FeeMetadata, IsmpDispatcher},
+	host::StateMachine,
+	messaging::Proof,
+};
+use pallet_ismp::dispatcher::{Message, WithdrawalRequest, HYPERBRIDGE_MODULE_ID};
+use pallet_ismp_host_executive::{HostParam, HostParams, WithdrawalParams};
 use polkadot_sdk::*;
-use sp_core::H256;
+use sp_core::{H256, U256};
+use sp_runtime::AccountId32;
 
 #[derive(
 	Debug, Clone, Encode, Decode, DecodeWithMemTracking, scale_info::TypeInfo, PartialEq, Eq,
@@ -37,13 +61,6 @@ pub struct WithdrawalProof {
 	pub beneficiary_details: Option<(Vec<u8>, Signature)>,
 }
 
-#[derive(RlpDecodable, RlpEncodable, Debug, Clone)]
-#[rlp(trailing)]
-pub struct FeeMetadata {
-	pub fee: alloy_primitives::U256,
-	pub sender: Address,
-}
-
 #[derive(
 	Debug, Clone, Encode, Decode, DecodeWithMemTracking, scale_info::TypeInfo, PartialEq, Eq,
 )]
@@ -54,4 +71,128 @@ pub struct WithdrawalInputData {
 	pub dest_chain: StateMachine,
 	/// The withdrawal beneficiary
 	pub beneficiary: Option<Vec<u8>>,
+}
+
+impl<T: Config> Pallet<T>
+where
+	<T as frame_system::Config>::Hash: From<H256>,
+	<T as frame_system::Config>::AccountId: From<[u8; 32]>,
+	T::Balance: Into<u128>,
+{
+	pub fn withdraw(withdrawal_data: WithdrawalInputData) -> DispatchResult {
+		let address = match &withdrawal_data.signature {
+			Signature::Evm { address, .. } => address.clone(),
+			Signature::Sr25519 { public_key, .. } => public_key.clone(),
+			Signature::Ed25519 { public_key, .. } => public_key.clone(),
+		};
+
+		let nonce = Nonce::<T>::get(address.clone(), withdrawal_data.dest_chain);
+		let msg = message(nonce, withdrawal_data.dest_chain, withdrawal_data.beneficiary.clone());
+
+		match &withdrawal_data.signature {
+			Signature::Evm { address, .. } => {
+				let eth_address = withdrawal_data
+					.signature
+					.verify(&msg, None)
+					.map_err(|_| Error::<T>::InvalidSignature)?;
+				if &eth_address != address {
+					Err(Error::<T>::InvalidPublicKey)?
+				}
+			},
+			Signature::Sr25519 { .. } => {
+				// Verify signature with public key provided in signature enum
+				withdrawal_data
+					.signature
+					.verify(&msg, None)
+					.map_err(|_| Error::<T>::InvalidSignature)?;
+			},
+			Signature::Ed25519 { .. } => {
+				// Verify signature with public key provided in signature enum
+				withdrawal_data
+					.signature
+					.verify(&msg, None)
+					.map_err(|_| Error::<T>::InvalidSignature)?;
+			},
+		};
+		let available_amount = Fees::<T>::get(withdrawal_data.dest_chain, address.clone());
+
+		if available_amount <
+			Self::min_withdrawal_amount(withdrawal_data.dest_chain)
+				.unwrap_or(MinWithdrawal::get())
+		{
+			Err(Error::<T>::NotEnoughBalance)?
+		}
+
+		let dispatcher = <T as Config>::IsmpHost::default();
+
+		Nonce::<T>::try_mutate(address.clone(), withdrawal_data.dest_chain, |value| {
+			*value += 1;
+			Ok::<(), ()>(())
+		})
+		.map_err(|_| Error::<T>::ErrorCompletingCall)?;
+
+		let beneficiary_address = withdrawal_data.beneficiary.clone().unwrap_or(address.clone());
+		let (to, body) = match withdrawal_data.dest_chain {
+			s if s.is_substrate() => (
+				HYPERBRIDGE_MODULE_ID.to_vec(),
+				Message::WithdrawRelayerFees(WithdrawalRequest {
+					amount: available_amount.low_u128(),
+					account: AccountId32::try_from(&beneficiary_address[..])
+						.map_err(|_| Error::<T>::InvalidPublicKey)?,
+				})
+				.encode(),
+			),
+			_ => {
+				let HostParam::EvmHostParam(params) =
+					HostParams::<T>::get(withdrawal_data.dest_chain)
+						.ok_or_else(|| Error::<T>::MissingMangerAddress)?;
+
+				let body = WithdrawalParams {
+					beneficiary_address: beneficiary_address.clone(),
+					amount: available_amount.into(),
+					token: params.fee_token,
+				}
+				.abi_encode()
+				.map_err(|_| Error::<T>::InvalidPublicKey)?;
+
+				(params.host_manager.0.to_vec(), body)
+			},
+		};
+
+		let post = DispatchPost {
+			dest: withdrawal_data.dest_chain,
+			from: MODULE_ID.to_vec(),
+			to,
+			body,
+			timeout: 0,
+		};
+
+		// Account is not useful in this case
+		dispatcher
+			.dispatch_request(
+				DispatchRequest::Post(post),
+				FeeMetadata { payer: [0u8; 32].into(), fee: Default::default() },
+			)
+			.map_err(|_| Error::<T>::DispatchFailed)?;
+
+		Fees::<T>::insert(withdrawal_data.dest_chain, address.clone(), U256::zero());
+
+		Self::deposit_event(Event::<T>::Withdraw {
+			address: sp_runtime::BoundedVec::truncate_from(address.clone()),
+			beneficiary_address: sp_runtime::BoundedVec::truncate_from(beneficiary_address),
+			state_machine: withdrawal_data.dest_chain,
+			amount: available_amount,
+		});
+
+		Ok(())
+	}
+}
+
+/// Signed payload for [`WithdrawalInputData`]. Includes the per-relayer nonce so a captured
+/// signature cannot be replayed.
+pub fn message(nonce: u64, dest_chain: StateMachine, beneficiary: Option<Vec<u8>>) -> [u8; 32] {
+	if let Some(beneficiary) = beneficiary {
+		return sp_io::hashing::keccak_256(&(nonce, dest_chain, beneficiary).encode());
+	}
+	sp_io::hashing::keccak_256(&(nonce, dest_chain).encode())
 }
